+++ R/table_functions.R
@@ -287,6 +287,14 @@ assemble_fu_allocation_tables <- function(incomplete_allocation_tables,
 #' Note that this function can accept tidy or wide by year data frames.
 #' The return value is a tidy data frame.
 #'
+#' Note that the `.values` argument applies for both
+#' `incomplete_eta_fu_tables` and
+#' `completed_fu_allocation_tables`.
+#' Callers should ensure that value columns in both
+#' data frames (`incomplete_eta_fu_tables` and `completed_fu_allocation_tables`)
+#' are named identically and that name is passed into the
+#' `.values` argument.
+#'
 #' @param incomplete_eta_fu_tables An incomplete data frame of final-to-useful efficiencies for all Machines in `completed_fu_allocation_tables`.
 #' @param exemplar_lists A data frame containing `country` and `year` columns along with a column of ordered vectors of strings
 #'                       telling which countries should be considered exemplars for the country and year of this row.
@@ -297,9 +305,9 @@ assemble_fu_allocation_tables <- function(incomplete_allocation_tables,
 #' @param which_quantity A vector of quantities to be completed in the eta_FU table.
 #'                       Default is `c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u)`.
 #'                       Must be one or both of the default values.
-#' @param country,year See `IEATools::iea_cols`.
-#' @param exemplars,exemplar_tables,alloc_data,incomplete_eta_tables,complete_eta_tables
-#'                    See `SEAPSUTWorkflows::exemplar_names`.
+#' @param country,method,energy_type,last_stage,year,unit,e_dot See `IEATools::iea_cols`.
+#' @param machine,eu_product,eta_fu,phi_u,c_source,eta_fu_phi_u_source,e_dot_machine,e_dot_machine_perc,quantity,maximum_values,e_dot_perc,.values See `IEATools::template_cols`.
+#' @param exemplars,exemplar_tables,alloc_data,incomplete_eta_tables,complete_eta_tables See `SEAPSUTWorkflows::exemplar_names`.
 #'
 #' @return A tidy data frame containing completed final-to-useful efficiency tables.
 #'
@@ -307,7 +315,7 @@ assemble_fu_allocation_tables <- function(incomplete_allocation_tables,
 #'
 #' @examples
 #' # Make some incomplete efficiency tables for GHA by removing Wood cookstoves.
-#' # Information from the exemplar, ZAF, will supply efficiency for Wood cookstoves.
+#' # Information from the exemplar, ZAF, will supply efficiency for Wood cookstoves for GHA.
 #' incomplete_eta_fu_tables <- IEATools::load_eta_fu_data() %>%
 #'   dplyr::filter(! (Country == "GHA" & Machine == "Wood cookstoves"))
 #' # The rows for Wood cookstoves are missing.
@@ -336,18 +344,50 @@ assemble_eta_fu_tables <- function(incomplete_eta_fu_tables,
                                    max_year = NULL,
                                    which_quantity = c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u),
                                    country = IEATools::iea_cols$country,
+                                   method = IEATools::iea_cols$method,
+                                   energy_type = IEATools::iea_cols$energy_type,
+                                   last_stage = IEATools::iea_cols$last_stage,
+                                   unit = IEATools::iea_cols$unit,
                                    year = IEATools::iea_cols$year,
+                                   e_dot = IEATools::iea_cols$e_dot,
+
+                                   machine = IEATools::template_cols$machine,
+                                   eu_product = IEATools::template_cols$eu_product,
+                                   eta_fu = IEATools::template_cols$eta_fu,
+                                   phi_u = IEATools::template_cols$phi_u,
+                                   c_source = IEATools::template_cols$c_source,
+                                   eta_fu_phi_u_source = IEATools::template_cols$eta_fu_phi_u_source,
+                                   e_dot_machine = IEATools::template_cols$e_dot_machine,
+                                   e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
+                                   quantity = IEATools::template_cols$quantity,
+                                   maximum_values = IEATools::template_cols$maximum_values,
+                                   e_dot_perc = IEATools::template_cols$e_dot_perc,
+
                                    exemplars = SEAPSUTWorkflow::exemplar_names$exemplars,
                                    exemplar_tables = SEAPSUTWorkflow::exemplar_names$exemplar_tables,
                                    alloc_data = SEAPSUTWorkflow::exemplar_names$alloc_data,
                                    incomplete_eta_tables = SEAPSUTWorkflow::exemplar_names$incomplete_eta_table,
-                                   complete_eta_tables = SEAPSUTWorkflow::exemplar_names$complete_eta_table) {
+                                   complete_eta_tables = SEAPSUTWorkflow::exemplar_names$complete_eta_table,
+
+                                   .values = IEATools::template_cols$.values) {
 
   which_quantity <- match.arg(which_quantity, several.ok = TRUE)
 
   # The FU allocation tables and the incomplete efficiency tables are easier to deal with when they are tidy.
-  tidy_incomplete_eta_fu_tables <- IEATools::tidy_eta_fu_table(incomplete_eta_fu_tables)
-  tidy_allocation_tables <- IEATools::tidy_fu_allocation_table(completed_fu_allocation_tables)
+  tidy_incomplete_eta_fu_tables <- IEATools::tidy_eta_fu_table(incomplete_eta_fu_tables,
+                                                               year = year,
+                                                               e_dot_machine = e_dot_machine,
+                                                               e_dot_machine_perc = e_dot_machine_perc,
+                                                               quantity = quantity,
+                                                               maximum_values = maximum_values,
+                                                               .values = .values)
+  tidy_allocation_tables <- IEATools::tidy_fu_allocation_table(completed_fu_allocation_tables,
+                                                               year = year,
+                                                               e_dot = e_dot,
+                                                               e_dot_perc = e_dot_perc,
+                                                               quantity = quantity,
+                                                               maximum_values = maximum_values,
+                                                               .values = .values)
   if (!is.null(max_year)) {
     tidy_incomplete_eta_fu_tables <- tidy_incomplete_eta_fu_tables %>%
       dplyr::filter(.data[[year]] <= max_year)
@@ -392,7 +432,27 @@ assemble_eta_fu_tables <- function(incomplete_eta_fu_tables,
                                        eta_fu_table = .data[[incomplete_eta_tables]],
                                        exemplar_eta_fu_tables = .data[[exemplar_tables]],
                                        fu_allocation_table = .data[[alloc_data]],
-                                       which_quantity = list(which_quantity))
+                                       which_quantity = list(which_quantity),
+
+                                       country = country,
+                                       method = method,
+                                       energy_type = energy_type,
+                                       last_stage = last_stage,
+                                       e_dot = e_dot,
+                                       unit = unit,
+                                       year = year,
+                                       machine = machine,
+                                       eu_product = eu_product,
+                                       e_dot_perc = e_dot_perc,
+                                       e_dot_machine = e_dot_machine,
+                                       e_dot_machine_perc = e_dot_machine_perc,
+                                       eta_fu = eta_fu,
+                                       phi_u = phi_u,
+                                       quantity = quantity,
+                                       maximum_values = maximum_values,
+                                       c_source = c_source,
+                                       eta_fu_phi_u_source = eta_fu_phi_u_source,
+                                       .values = .values)
       )
   }) %>%
     dplyr::bind_rows()